Refactor: InvitationServiceTest 통합 (#46)
