+++ src/test/java/com/livable/server/invitation/service/InvitationServiceTest.java
@@ -13,8 +13,12 @@ import com.livable.server.invitation.repository.OfficeRepository;
 import com.livable.server.invitation.service.data.InvitationBasicData;
 import com.livable.server.member.repository.MemberRepository;
 import com.livable.server.reservation.repository.ReservationRepository;
+import com.livable.server.visitation.domain.VisitationErrorCode;
+import com.livable.server.visitation.dto.VisitationResponse;
+import com.livable.server.visitation.mock.MockInvitationDetailTimeDto;
 import com.livable.server.visitation.repository.ParkingLogRepository;
 import com.livable.server.visitation.repository.VisitorRepository;
+import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -36,6 +40,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.then;
+import static org.mockito.Mockito.times;
 
 @ExtendWith(MockitoExtension.class)
 class InvitationServiceTest {
@@ -941,6 +947,47 @@ class InvitationServiceTest {
                         "공용 B"
                 )
         ));
+    }
+
+    @DisplayName("InvitationService.findInvitationTime 성공 테스트")
+    @Test
+    void findInvitationTimeSuccessTest() {
+
+        // Given
+        MockInvitationDetailTimeDto mockInvitationDetailTimeDto = new MockInvitationDetailTimeDto();
+        Invitation invitation = Invitation.builder()
+                .startTime(mockInvitationDetailTimeDto.getStartTime())
+                .endTime(mockInvitationDetailTimeDto.getEndTime())
+                .startDate(mockInvitationDetailTimeDto.getStartDate())
+                .endDate(mockInvitationDetailTimeDto.getEndDate())
+                .build();
+
+        given(invitationRepository.findInvitationDetailTimeByVisitorId(anyLong()))
+                .willReturn(Optional.of(mockInvitationDetailTimeDto));
+
+        // When
+        VisitationResponse.InvitationTimeDto invitationTime = invitationService.findInvitationTime(1L);
+
+        // Then
+        then(invitationRepository).should(times(1)).findInvitationDetailTimeByVisitorId(anyLong());
+        AssertionsForClassTypes.assertThat(invitationTime).usingRecursiveComparison().isEqualTo(invitation);
+    }
+
+    @DisplayName("InvitationService.findInvitationTime 실패 테스트")
+    @Test
+    void findInvitationTimeFailTest() {
+
+        // Given
+        given(invitationRepository.findInvitationDetailTimeByVisitorId(anyLong())).willReturn(Optional.empty());
+
+        // When
+        GlobalRuntimeException globalRuntimeException = assertThrows(
+                GlobalRuntimeException.class, () -> invitationService.findInvitationTime(anyLong())
+        );
+
+        // Then
+        AssertionsForClassTypes.assertThat(globalRuntimeException.getErrorCode()).isEqualTo(VisitationErrorCode.NOT_FOUND);
+        then(invitationRepository).should(times(1)).findInvitationDetailTimeByVisitorId(anyLong());
     }
 
 }
